增加用户提现接口

+++ src/main/java/xyz/nadev/house/entity/User.java
@@ -1,13 +1,14 @@
 package xyz.nadev.house.entity;
 
 import lombok.Data;
-
-import javax.persistence.*;
 import org.hibernate.annotations.DynamicInsert;
 import org.hibernate.annotations.DynamicUpdate;
 import org.springframework.data.annotation.CreatedDate;
 import org.springframework.data.annotation.LastModifiedDate;
 import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+import javax.persistence.*;
+import java.math.BigDecimal;
 
 /**
  * 用户信息表
@@ -83,7 +84,7 @@ public class User {
 	 * default value: 0.00
 	 */
 	@Column(name = "money", nullable = false)
-	private Double money;
+	private BigDecimal money;
 
 	/**
 	 * 创建时间
